separate otb provider into a plugin
