+++ python/plugins/otbprovider/metadata.txt
@@ -1,0 +1,17 @@
+[general]
+name=OrfeoToolbox provider
+description=OrfeoToolbox Processing provider
+about=OrfeoToolbox Processing provider
+category=Analysis
+version=2.12.99
+qgisMinimumVersion=3.0
+
+author=Victor Olaya
+
+icon=:/images/themes/default/providerOtb.svg
+
+homepage=https://qgis.org
+tracker=https://github.com/qgis/QGIS/issues
+repository=https://github.com/qgis/QGIS
+
+hasProcessingProvider=yes
