+++ python/plugins/processing/algs/CMakeLists.txt
@@ -2,7 +2,6 @@ file(GLOB PY_FILES *.py)
 
 add_subdirectory(help)
 add_subdirectory(gdal)
-add_subdirectory(otb)
 add_subdirectory(qgis)
 
 PLUGIN_INSTALL(processing algs ${PY_FILES})
