--- python/plugins/processing/algs/otb/CMakeLists.txt
@@ -1,2 +1,0 @@
-file(GLOB PY_FILES *.py)
-PLUGIN_INSTALL(processing algs/otb ${PY_FILES})
